Lighting, Blocks, and Chunks
The first thing I added in this version is some simple point lighting.
However, it does not work as it should due to the fact that I am using
8-vertex cubes.
I started on a management system for the worlds and started with an
OctTree system. After a little more thought on it and some research I
switched to a vector-based chunk system that is currently in progress.